+++ src/resources/widgets/user/index/table.widget.php
@@ -9,16 +9,6 @@ return [
     'options' => [
         'hover' => true,
         'columns' => [
-            'id' => [
-                'title' => '&nbsp;',
-                'valign' => 'middle',
-                'widget' => [
-                    'type' => 'raw_html',
-                    'html' => function ($issue) {
-                        return '<input type="checkbox" name="issue_' . $issue->id . '" />';
-                    },
-                ],
-            ],
             'avatar' => [
                 'widget' => [
                     'type' => 'avatar',